Changes to bel_dyn_onestep
The latest version of the code contains an update in bel_dyn_onestep, line 22. Instead of taking into account all opinions held in memory for the calculation of the individually evolving affinity threshold, the new version only takes into account the current time-step when calculating the individually evolving affinity threshold.